backend/requirements.txt
asgiref==3.7.2
dj-rest-auth==2.1.11
Django==4.2.6
django-cors-headers==3.10.1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
